runtime: Refactor SGX DCAP quote verification module
